Basic code to handle empty category and tags setting
Added most of the code; still need to test to see if an empty string is accepted by the WP API.

Also need to check the format in the config options. This is a bit of scope creep but while I'm working on this it makes sense to add it.