Sorted out issue w/search API response
The search API returns a list of matches so we just need [0]. Some minor cosmetic changes made it in as well.

Now the only remaining broken thing is posts for weeks that end in the future, which need to have a post date of the current day or they won't show up. Which is not all that bad but let's fix it anyhow.