Fix for movies without ratings
It's not a problem for fetchcsv, since if there's not a rating there, the CSV library will just fork over a null for the rating. However, the property won't exist in the RSS feed if you didn't rate the movie.

At some point possibly I'll build a test feed; the movie that triggered this problem for me is over 20 entries back in my queue so I'm not 100% sure of the fix. However, it's still processing movies with ratings fine and the code fix was really simple.